some clean up in net
